id numerico

+++ api-vakita/ms-payment/src/main/java/grupo3/mspayment/repository/ICreditCardMongoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ICreditCardMongoRepository extends MongoRepository<CreditCard, Long> {
+public interface ICreditCardMongoRepository extends MongoRepository<CreditCard, Integer> {
 }
